+++ libraries/repro-curl/src/urlencode.cpp
@@ -10,7 +10,7 @@ std::string Urlencode::decode(const std::string& s)
     return decode( s.c_str(), s.size() );
 }
 
-std::string Urlencode::decode(const char* s, int len)
+std::string Urlencode::decode(const char* s, size_t len)
 {
 	std::ostringstream oss;
 	for( int i = 0; i < len; i++) {
@@ -23,7 +23,7 @@ std::string Urlencode::decode(const char* s, int len)
 	}
 	std::string tmp = oss.str();
 
-    char* c = curl_unescape( tmp.c_str() , tmp.size() );
+    char* c = curl_unescape( tmp.c_str() , (int) tmp.size() );
     std::string result(c);
     curl_free(c);
     return result;
@@ -34,10 +34,10 @@ std::string Urlencode::encode(const std::string& s)
     return encode( s.c_str(), s.size() );
 }
 
-std::string Urlencode::encode(const char* s, int len)
+std::string Urlencode::encode(const char* s, size_t len)
 {
     CURL* curl = curl_easy_init();
-    char* c = curl_easy_escape( curl , s , len );
+    char* c = curl_easy_escape( curl , s , (int)len );
     std::string result(c);
     curl_free(c);
     curl_easy_cleanup(curl);
